Better exclusion of outer face.
The "outer face", that is, the face that consists of all outer walls,
has been excluded until now by virtue of the safety_counter - basically
it'll have more walls than the maximum per cell and is thus excluded.
That's an ugly hack.  I left the safety counter in for safety, but now
the exclusion happens because the walls on that face are
counter-clockwise since the face is really "inside out".