Added "round" maze style.
In this style, xsize is the number of "rings" and ysize is the density.
The density divided by 2 is the number of cells on the inner ring.  As
rings are added to the maze, the density is doubled whenever the arc
length of the outer wall is more than double the arc length of the
outer wall of the first ring.

It's likely that there's a better way to do that with a Fibonacci
sequence, I just haven't given it much thought yet.